feat(catalog): Expanded catalog structure for evolving logic
