+++ Interfaces/TenantAreaInterface/catalog/index.ts
@@ -1,0 +1,2 @@
+export * from "./queries";
+export * from "./selection";
